PKUmoocServer/requirements.txt
asgiref==3.7.2
sqlparse==0.4.4
Django==4.2.7
pytz==2023.3.post1
djangorestframework==3.14.0
httpie
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
